+++ src/Ecomprocessing/Controllers/Frontend/EcomprocessingCheckoutPayment.php
@@ -51,11 +51,13 @@ class Shopware_Controllers_Frontend_EcomprocessingCheckoutPayment extends Fronte
     public function payAction()
     {
         $tokenizationHelper = $this->container->get('ecomprocessing.wpf_tokenization_service');
+
         // Process request
         try {
             // Load the Payment Data
             $paymentData = new PaymentData($this->Request()->getParams());
             $paymentData->setOrderBasket($this->getBasket());
+            $paymentData->setUser($this->getUser());
 
             // Log the Payment Data
             $this->logger->info('Payment Data', $this->getPaymentShortName(), $paymentData->toArray());