+++ src/Ecomprocessing/Resources/views/backend/ecomprocessing_payment_methods/store/ecomprocessing_checkout_sca_exemption_amount_store.js
@@ -1,0 +1,13 @@
+// Ecomprocessing SCA Exemption option amount
+
+
+
+
+Ext.define('Shopware.apps.EcomprocessingPaymentMethods.store.EcomprocessingCheckoutScaExemptionAmountStore', {
+    extend: 'Ext.data.Store',
+
+    fields: [
+        { name: 'option', type: 'string' },
+        { name: 'value', type: 'int' }
+    ]
+});